Move VirtualButtonState and Timing into a more appropriate location
