+++ src/action_state.rs
@@ -1,143 +1,12 @@
 //! This module contains [`ActionState`] and its supporting methods and impls.
 
-use crate::prelude::UserInput;
+use crate::buttonlike_user_input::{Timing, UserInput, VirtualButtonState};
 use crate::Actionlike;
+
 use bevy_ecs::{component::Component, entity::Entity};
 use bevy_utils::{Duration, Instant};
 use serde::{Deserialize, Serialize};
 use std::marker::PhantomData;
-
-/// The current state of a particular virtual button,
-/// corresponding to a single [`Actionlike`] action.
-///
-/// Detailed timing information for the button can be accessed through the stored [`Timing`] value
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub enum VirtualButtonState {
-    /// This button is currently pressed
-    Pressed(Timing, Vec<UserInput>),
-    /// This button is currently released
-    Released(Timing),
-}
-
-/// Stores the timing information for a [`VirtualButtonState`]
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct Timing {
-    /// The [`Instant`] at which the button was pressed or released
-    ///
-    /// Recorded as the [`Time`](bevy::core::Time) at the start of the tick after the state last changed.
-    /// If this is none, [`ActionState::update`] has not been called yet.
-    #[serde(skip)]
-    pub instant_started: Option<Instant>,
-    /// The [`Duration`] for which the button has been pressed or released.
-    ///
-    /// This begins at [`Duration::ZERO`] when [`ActionState::update`] is called.
-    pub current_duration: Duration,
-    /// The [`Duration`] for which the button was pressed or released before the state last changed.
-    pub previous_duration: Duration,
-}
-
-impl PartialOrd for Timing {
-    fn partial_cmp(&self, other: &Self) -> Option<std::cmp::Ordering> {
-        self.current_duration.partial_cmp(&other.current_duration)
-    }
-}
-
-impl PartialOrd for VirtualButtonState {
-    fn partial_cmp(&self, other: &Self) -> Option<std::cmp::Ordering> {
-        match self {
-            VirtualButtonState::Pressed(timing, _) => match other {
-                VirtualButtonState::Pressed(other_timing, _) => timing.partial_cmp(other_timing),
-                VirtualButtonState::Released(other_timing) => timing.partial_cmp(other_timing),
-            },
-            VirtualButtonState::Released(timing) => match other {
-                VirtualButtonState::Pressed(other_timing, _) => timing.partial_cmp(other_timing),
-                VirtualButtonState::Released(other_timing) => timing.partial_cmp(other_timing),
-            },
-        }
-    }
-}
-
-impl VirtualButtonState {
-    /// Is the button currently pressed?
-    #[inline]
-    #[must_use]
-    pub fn pressed(&self) -> bool {
-        match self {
-            VirtualButtonState::Pressed(_, _) => true,
-            VirtualButtonState::Released(_) => false,
-        }
-    }
-
-    /// Is the button currently released?
-    #[inline]
-    #[must_use]
-    pub fn released(&self) -> bool {
-        match self {
-            VirtualButtonState::Pressed(_, _) => false,
-            VirtualButtonState::Released(_) => true,
-        }
-    }
-
-    /// Was the button pressed since the last time [`ActionState::update`] was called?
-    #[inline]
-    #[must_use]
-    pub fn just_pressed(&self) -> bool {
-        match self {
-            VirtualButtonState::Pressed(timing, _) => timing.instant_started.is_none(),
-            VirtualButtonState::Released(_timing) => false,
-        }
-    }
-
-    /// Was the button released since the last time [`ActionState::update`] was called?
-    #[inline]
-    #[must_use]
-    pub fn just_released(&self) -> bool {
-        match self {
-            VirtualButtonState::Pressed(_timing, _) => false,
-            VirtualButtonState::Released(timing) => timing.instant_started.is_none(),
-        }
-    }
-
-    /// The [`Instant`] at which the button was pressed or released
-    ///
-    /// Recorded as the [`Time`](bevy::core::Time) at the start of the tick after the state last changed.
-    /// If this is none, [`ActionState::update`] has not been called yet.
-    #[inline]
-    #[must_use]
-    pub fn instant_started(&self) -> Option<Instant> {
-        match self {
-            VirtualButtonState::Pressed(timing, _) => timing.instant_started,
-            VirtualButtonState::Released(timing) => timing.instant_started,
-        }
-    }
-
-    /// The [`Duration`] for which the button has been pressed or released.
-    ///
-    /// This begins at [`Duration::ZERO`] when [`ActionState::update`] is called.
-    #[inline]
-    #[must_use]
-    pub fn current_duration(&self) -> Duration {
-        match self {
-            VirtualButtonState::Pressed(timing, _) => timing.current_duration,
-            VirtualButtonState::Released(timing) => timing.current_duration,
-        }
-    }
-    /// The [`Duration`] for which the button was pressed or released before the state last changed.
-    #[inline]
-    #[must_use]
-    pub fn previous_duration(&self) -> Duration {
-        match self {
-            VirtualButtonState::Pressed(timing, _) => timing.previous_duration,
-            VirtualButtonState::Released(timing) => timing.previous_duration,
-        }
-    }
-}
-
-impl Default for VirtualButtonState {
-    fn default() -> Self {
-        VirtualButtonState::Released(Timing::default())
-    }
-}
 
 /// Stores the canonical input-method-agnostic representation of the inputs received
 ///
